+++ test/partial_reindex_test.rb
@@ -1,7 +1,7 @@
 require_relative "test_helper"
 
 class PartialReindexTest < Minitest::Test
-  def test_class_method
+  def test_relation_inline
     store [{name: "Hi", color: "Blue"}]
 
     # normal search
@@ -15,7 +15,7 @@ class PartialReindexTest < Minitest::Test
     Searchkick.callbacks(false) do
       product.save!
     end
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
 
     # index not updated
     assert_search "hi", ["Hi"], fields: [:name], load: false
@@ -29,7 +29,7 @@ class PartialReindexTest < Minitest::Test
     assert_search "blue", ["Bye"], fields: [:color], load: false
   end
 
-  def test_instance_method
+  def test_record_inline
     store [{name: "Hi", color: "Blue"}]
 
     # normal search
@@ -43,7 +43,7 @@ class PartialReindexTest < Minitest::Test
     Searchkick.callbacks(false) do
       product.save!
     end
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
 
     # index not updated
     assert_search "hi", ["Hi"], fields: [:name], load: false
@@ -56,8 +56,32 @@ class PartialReindexTest < Minitest::Test
     assert_search "blue", ["Bye"], fields: [:color], load: false
   end
 
-  def test_instance_method_async
+  def test_record_async
     product = Product.create!(name: "Hi")
     product.reindex(:search_data, mode: :async)
+  end
+
+  def test_relation_missing
+    store [{name: "Hi", color: "Blue"}]
+
+    product = Product.first
+    Product.searchkick_index.remove(product)
+
+    error = assert_raises(Searchkick::ImportError) do
+      Product.reindex(:search_name)
+    end
+    assert_match "document missing", error.message
+  end
+
+  def test_record_missing
+    store [{name: "Hi", color: "Blue"}]
+
+    product = Product.first
+    Product.searchkick_index.remove(product)
+
+    error = assert_raises(Searchkick::ImportError) do
+      product.reindex(:search_name)
+    end
+    assert_match "document missing", error.message
   end
 end