+++ test/support/mongoid.rb
@@ -21,6 +21,10 @@ class Product
   field :longitude, type: BigDecimal
   field :description
   field :alt_description
+  field :embedding, type: Array
+  field :embedding2, type: Array
+  field :factors, type: Array
+  field :vector, type: Array
 end
 
 class Store